// src/store/effects/gql/graphql-types/DeleteCountry.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: DeleteCountry
// ====================================================

export interface DeleteCountry_deleteCountry {
  uid: string;
}

export interface DeleteCountry {
  deleteCountry: DeleteCountry_deleteCountry;
}

export interface DeleteCountryVariables {
  uid: string;
}